+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/TxGen.hs
@@ -917,7 +917,7 @@ genAlonzoTxAndInfo proof slot = do
     if withdrawalAmount == Coin 0
       then pure Nothing
       else Just . coreTxOut proof <$> genTxOut proof (inject withdrawalAmount)
-  let wdrlCreds = map (getRwdCred . fst) $ Map.toAscList wdrlMap
+  let wdrlCreds = map (raCredential . fst) $ Map.toAscList wdrlMap
   (IsValid v2, mkWithdrawalsWits) <-
     redeemerWitnessMaker proof Rewarding $ map (Just . (,) genDatum) wdrlCreds